CalibratioonTestServiceImplTest edited and moved to proper directory
